requirements.txt
censys
python-dotenv
mcp